Split the MEM type class into MemOps and MemSpec
This makes it easier to define new instances: define the operations
first, then prove things about them. More generally, it makes sense to
separate the code, which should be transparent and will be extracted,
from the associated proofs, which can be opaque and will be erased.

Nevertheless, for now MemSpec is still in Type rather than Prop because
some fields are informative (ie. use [sig] rather that [exists]).