Lift memory states along a comonad (near complete)
Given a comonad W, a type of memory states mem, and a way to lift the
empty memory state from mem to (W mem), we can lift the operations and
properties of Mem.MemSpec from mem to (W mem).

The proof for Mem.storebytes_split is missing but it should not be too
difficult to make it go through.