Work around 'intuition' destructing MemSpec instances.
The split of MEM into MemOps and MemSpec changes the behavior of the
'intuition' tactic, which now destructs instances of MemSpec. This
results in a huge number of hypotheses. Worse, it changes some
hypothesis names, and (at lease) common/Events.v breaks as a result.

Apparently, this happens because no two field of MemSpec depend on each
other any more. To work around this I reintroduce some dependency using
bogus fields in Mem.MemSpec.