Rename the type classes
I adopt the convention that since the "predicate" type classes represent
the whole thing, they get the unmodified name of the associated algebraic
structures (or a short description thereof), rather than adding a "Spec"
suffix as is currently done (cf. "MemSpec" for instance).  On the other
hand, the classes which contain only the operations keep their current
name as "FooOps".

Hence the names are modified as follows:
  * MemOps -> MemoryOps
  * MemSpec -> MemoryStates
  * ExtFunOps is unchanged
  * ExternalFunctions is introduced for consitency (but is 'unit')
  * ExtCallOps is unchanged
  * ExtCallSpec -> ExternalCalls

In addition, an instance of ExternalFunctions is added to AST.EFImpl, and
the code is modified to add an instance of ExternalFunctions to the
context rather than only ExtFunOps where appropriate.