Separate [builtin_function] from [external_function]
The type [external_function] is used for two different purposes:

  * by the [External] constructor of the [fundef] type uses, which uses
    it by contrast with [Internal] function whose body is given in the
    program being compiled;

  * directly by the [Ebuiltin]/[Sbuiltin]/[Pbuiltin] family of syntactic
    constructions, which are used for most of the compiler's built-in
    operations (which are, essentially, pre-defined pieces of inline
    assembly).

However, by and large, these two uses use disjoint portions of the
[external_function] type, and so we split those so that:

  * [External] still uses [external_function], which can still be
    parametrized in [CompilerConfiguration];

  * The [Xbuiltin] family of constructions use the new
    [builtin_function], which is not part of the configuration but a
    regular old constant type (although we do declare an instance of
    the [ExternalFunctions] type for [builtin_function], so that the
    expected theorems are still available for them).

There is one case where an element of [external_function] was used in
both of these contexts, however. Nameley, [EF_builtin] is read by the
parser as just a particular flavor of an actual external function,
provided that its name is recognized as one of the available compiler
builtin. However, calls to such external functions are converted by the
selection pass into [Sbuiltin] instructions, so as to be implemented,
essentially by expanding them into assembly instruction at the so-called
"pretty-printing" phase.

In order for Selection to be able to do this, we need to duplicate this
particular constructor of [external_function] as the [EF_inlined_builtin]
constructor of [builtin_function].